models/requirements.txt
pandas==2.2.3
python-dotenv==1.0.1
fpgrowth_py==1.0.0
